collations: extract the servenv dependency
